+++ linopy/solvers.py
@@ -24,7 +24,7 @@ from linopy.constants import (
     TerminationCondition,
 )
 
-quadratic_solvers = ["gurobi", "xpress", "cplex", "highs"]
+quadratic_solvers = ["gurobi", "xpress", "cplex", "highs", "mindopt"]
 
 available_solvers = []
 
@@ -54,11 +54,15 @@ with contextlib.suppress(ImportError):
     import xpress
 
     available_solvers.append("xpress")
+with contextlib.suppress(ImportError):
+    import mindoptpy
+
+    available_solvers.append("mindopt")
 logger = logging.getLogger(__name__)
 
 
 io_structure = dict(
-    lp_file={"gurobi", "xpress", "cbc", "glpk", "cplex"}, blocks={"pips"}
+    lp_file={"gurobi", "xpress", "cbc", "glpk", "cplex", "mindopt"}, blocks={"pips"}
 )
 
 
@@ -701,6 +705,83 @@ def run_xpress(
 
     solution = safe_get_solution(status, get_solver_solution)
     maybe_adjust_objective_sign(solution, model.objective.sense, io_api, "xpress")
+
+    return Result(status, solution, m)
+
+
+def run_mindopt(
+    model,
+    io_api=None,
+    problem_fn=None,
+    solution_fn=None,
+    log_fn=None,
+    warmstart_fn=None,
+    basis_fn=None,
+    keep_files=False,
+    env=None,
+    **solver_options,
+):
+    """
+    Solve a linear problem using the MindOpt solver.
+
+    https://solver.damo.alibaba.com/doc/en/html/index.html
+
+    For more information on solver options, see
+    https://solver.damo.alibaba.com/doc/en/html/API2/param/index.html
+    """
+    CONDITION_MAP = {
+        -1: "error",
+        0: "unknown",
+        1: "optimal",
+        2: "infeasible",
+        3: "unbounded",
+        4: "infeasible_or_unbounded",
+        5: "suboptimal",
+    }
+
+    if io_api is not None and io_api not in ["lp", "mps"]:
+        logger.warning(
+            f"IO setting '{io_api}' not available for mindopt solver. "
+            "Falling back to `lp`."
+        )
+
+    problem_fn = model.to_file(problem_fn)
+
+    problem_fn = maybe_convert_path(problem_fn)
+    log_fn = "" if not log_fn else maybe_convert_path(log_fn)
+
+    env = mindoptpy.Env(log_fn)
+    env.start()
+
+    m = mindoptpy.read(problem_fn, env)
+
+    for k, v in solver_options.items():
+        m.setParam(k, v)
+
+    m.optimize()
+
+    condition = m.status
+    termination_condition = CONDITION_MAP.get(condition, condition)
+    status = Status.from_termination_condition(termination_condition)
+    status.legacy_status = condition
+
+    def get_solver_solution() -> Solution:
+        objective = m.objval
+
+        sol = pd.Series({v.varname: v.X for v in m.getVars()}, dtype=float)
+        sol = set_int_index(sol)
+
+        try:
+            dual = pd.Series({c.constrname: c.DualSoln for c in m.getConstrs()})
+            dual = set_int_index(dual)
+        except mindoptpy.MindoptError:
+            logger.warning("Dual values of MILP couldn't be parsed")
+            dual = pd.Series(dtype=float)
+
+        return Solution(sol, dual, objective)
+
+    solution = safe_get_solution(status, get_solver_solution)
+    maybe_adjust_objective_sign(solution, model.objective.sense, io_api, "mindopt")
 
     return Result(status, solution, m)